Move function to 1D as 2D is not yet supported
